requirements.txt
# FastAPI and web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlmodel==0.0.16
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Geospatial
geoalchemy2==0.14.7
shapely==2.0.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Async and workers
celery==5.3.6
redis==5.0.3

# External services
openai==1.12.0
sendgrid==6.11.0
httpx==0.27.0

# Document processing
unstructured[pdf]==0.12.4
weasyprint==61.2

# Configuration and logging
pydantic-settings==2.2.1
structlog==24.1.0
python-json-logger==2.0.7
pyyaml==6.0.1

# Template rendering
jinja2==3.1.3

# Utilities
python-dotenv==1.0.1
slowapi==0.1.9

# Development and testing (optional)
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
ruff==0.3.3
mypy==1.9.0
black==24.3.0

# Production monitoring
sentry-sdk[fastapi]==1.42.0